+++ shared/packages/worker/src/worker/accessorHandlers/lib/CachedQuantelGateway.ts
@@ -1,18 +1,15 @@
 import { ClipSearchQuery, QuantelGateway } from 'tv-automation-quantel-gateway-client'
-import {
-	ClipData,
-	ClipDataSummary,
-	ConnectionDetails,
-	ServerInfo,
-	ZoneInfo,
-} from 'tv-automation-quantel-gateway-client/dist/quantelTypes'
+import { ClipDataSummary, ServerInfo, ZoneInfo } from 'tv-automation-quantel-gateway-client/dist/quantelTypes'
 
 const DEFAULT_CACHE_EXPIRE = 3000
-const PURGE_CACHE_INTERVAL = 60 * 1000
 
+/**
+ * This is a wrapper for the QuantelGateway class, adding a caching-layer to
+ */
 export class CachedQuantelGateway extends QuantelGateway {
 	private readonly _cache: Map<string, CacheEntry> = new Map()
 	private cacheExpire: number
+	private purgeExpiredCacheTimeout: NodeJS.Timeout | null = null
 
 	constructor(
 		config?:
@@ -25,68 +22,62 @@ export class CachedQuantelGateway extends QuantelGateway {
 	) {
 		super(config)
 		this.cacheExpire = config?.cacheExpire ?? DEFAULT_CACHE_EXPIRE
-		setInterval(() => this.purgeCache(), PURGE_CACHE_INTERVAL)
 	}
-	async connectToISA(ISAUrls: string | string[]): Promise<ConnectionDetails> {
-		const result = this.queryCache<ConnectionDetails>('connectToISA', [ISAUrls])
-		if (result !== undefined) return result
 
-		return this.ensureInCache('connectToISA', [ISAUrls], super.connectToISA(ISAUrls))
+	async purgeCacheSearchClip(searchQuery: ClipSearchQuery): Promise<ClipDataSummary[]> {
+		return this.clearCache('searchClip', [searchQuery]) ?? []
 	}
-	async getClip(clipId: number): Promise<ClipData | null> {
-		const result = this.queryCache<ClipData>('getClip', [clipId])
-		if (result !== undefined) return result
-
-		return this.ensureInCache('getClip', [clipId], super.getClip(clipId))
-	}
+	// searchClip() is used often. By caching it, we reduce the load on the server.
 	async searchClip(searchQuery: ClipSearchQuery): Promise<ClipDataSummary[]> {
-		const result = this.queryCache<ClipDataSummary[]>('searchClip', [searchQuery])
-		if (result !== undefined) return result
-
-		return this.ensureInCache('searchClip', [searchQuery], super.searchClip(searchQuery))
+		return this.ensureInCache('searchClip', [searchQuery], async () => super.searchClip(searchQuery))
 	}
 	async getZones(): Promise<ZoneInfo[]> {
-		const result = this.queryCache<ZoneInfo[]>('getZones', [])
-		if (result !== undefined) return result
-
-		return this.ensureInCache('getZones', [], super.getZones())
+		return this.ensureInCache('getZones', [], async () => super.getZones())
 	}
+	// getServer() is used often. By caching it, we reduce the load on the server.
 	async getServer(disableCache?: boolean): Promise<ServerInfo | null> {
-		const result = this.queryCache<ServerInfo>('getServer', [disableCache])
-		if (result !== undefined) return result
-
-		return this.ensureInCache('getServer', [disableCache], super.getServer(disableCache))
+		return this.ensureInCache('getServer', [disableCache], async () => super.getServer(disableCache))
 	}
-	async getServers(zoneId?: string): Promise<ServerInfo[]> {
-		const result = this.queryCache<ServerInfo[]>('getServers', [zoneId])
-		if (result !== undefined) return result
-
-		return this.ensureInCache('getServers', [zoneId], super.getServers(zoneId))
+	private getCacheKey(method: string, args: any[]) {
+		return `${method}_${JSON.stringify(args)}`
 	}
-
-	private queryCache<T>(method: string, args: any[]): Promise<T> | undefined {
+	private async clearCache(method: string, args: any[]): Promise<any | undefined> {
 		const cacheKey = this.getCacheKey(method, args)
 
-		const inCache = this._cache.get(cacheKey)
+		const cached = this._cache.get(cacheKey)
+		this._cache.delete(cacheKey)
 
-		// cache miss
-		if (inCache === undefined) return undefined
+		return cached ? cached.promise : undefined
+	}
+	private async ensureInCache<T>(method: string, args: any[], getValueFcn: () => Promise<T>): Promise<T> {
+		this.triggerPurgeExpiredFromCache()
 
-		// cache stale
-		const entryExpires = inCache.timestamp + this.cacheExpire
-		if (entryExpires < Date.now()) {
-			this._cache.delete(cacheKey)
-			return undefined
-		}
+		const cacheKey = this.getCacheKey(method, args)
 
-		if (inCache.result.state === 'rejected') {
-			return Promise.reject(inCache.result.err)
+		const cached = this._cache.get(cacheKey)
+		if (cached && Date.now() - cached.timestamp < this.cacheExpire) {
+			return cached.promise
 		} else {
-			return Promise.resolve(inCache.result.value)
+			const promise: Promise<any> = getValueFcn()
+
+			this._cache.set(cacheKey, {
+				timestamp: Date.now(),
+				promise: promise,
+			})
+			return promise
 		}
 	}
 
-	private purgeCache() {
+	private triggerPurgeExpiredFromCache() {
+		// Schedule a purging of expired packages
+		if (!this.purgeExpiredCacheTimeout) {
+			this.purgeExpiredCacheTimeout = setTimeout(() => {
+				this.purgeExpiredCacheTimeout = null
+				this.purgeExpiredFromCache()
+			}, this.cacheExpire + 100)
+		}
+	}
+	private purgeExpiredFromCache() {
 		const expiredKeys: string[] = []
 		this._cache.forEach((value, key) => {
 			if (value.timestamp >= Date.now() - this.cacheExpire) return
@@ -97,46 +88,9 @@ export class CachedQuantelGateway extends QuantelGateway {
 			this._cache.delete(key)
 		}
 	}
-
-	private async ensureInCache<T>(method: string, args: any[], answer: Promise<T>): Promise<T> {
-		const cacheKey = this.getCacheKey(method, args)
-
-		try {
-			const result = await answer
-			this._cache.set(cacheKey, {
-				timestamp: Date.now(),
-				result: {
-					state: 'resolved',
-					value: result,
-				},
-			})
-			return result
-		} catch (e) {
-			this._cache.set(cacheKey, {
-				timestamp: Date.now(),
-				result: {
-					state: 'rejected',
-					err: e,
-				},
-			})
-			return Promise.reject(e)
-		}
-	}
-
-	private getCacheKey(method: string, args: any[]) {
-		return `${method}_${JSON.stringify(args)}`
-	}
 }
 
 interface CacheEntry {
 	timestamp: number
-	result:
-		| {
-				state: 'rejected'
-				err: any
-		  }
-		| {
-				state: 'resolved'
-				value: any
-		  }
+	promise: Promise<any>
 }