+++ tests/internal-tests/src/__tests__/issues.spec.ts
@@ -5,7 +5,7 @@ import { ExpectedPackageStatusAPI } from '@sofie-automation/shared-lib/dist/pack
 import { Expectation, literal } from '@sofie-package-manager/api'
 import type * as fsMockType from '../__mocks__/fs'
 import { prepareTestEnviromnent, TestEnviromnent } from './lib/setupEnv'
-import { waitTime } from './lib/lib'
+import { waitUntil, waitTime } from './lib/lib'
 import { getLocalSource, getLocalTarget } from './lib/containers'
 import { WorkerAgent } from '@sofie-package-manager/worker'
 jest.mock('fs')
@@ -49,32 +49,36 @@ describe('Handle unhappy paths', () => {
 			[getLocalTarget('target0', 'file0Target.mp4')]
 		)
 
-		await waitTime(env.WAIT_JOB_TIME)
-
-		// Expect the Expectation to be waiting:
-		expect(env.expectationStatuses['copy0']).toMatchObject({
-			actualVersionHash: null,
-			statusInfo: {
-				status: /new|waiting/,
-				statusReason: {
-					tech: /not able to access file/i,
+		await waitUntil(() => {
+			// Expect the Expectation to be waiting:
+			expect(env.expectationStatuses['copy0']).toMatchObject({
+				actualVersionHash: null,
+				statusInfo: {
+					status: /new|waiting/,
+					statusReason: {
+						tech: /not able to access file/i,
+					},
 				},
-			},
-		})
+			})
+		}, env.WAIT_JOB_TIME)
+
+		expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
+			ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.NOT_FOUND
+		)
 
 		// Now the file suddenly pops up:
 		fs.__mockSetFile('/sources/source0/file0Source.mp4', 1234)
 
-		await waitTime(env.WAIT_SCAN_TIME)
-		await waitTime(env.ERROR_WAIT_TIME)
-		await waitTime(env.WAIT_JOB_TIME)
+		// Wait for the job to complete:
+		await waitUntil(() => {
+			// Expect the copy to have completed by now:
+			expect(env.containerStatuses['target0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
+				ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
+			)
+		}, env.WAIT_SCAN_TIME + env.ERROR_WAIT_TIME + env.WAIT_JOB_TIME)
 
-		// Expect the copy to have completed by now:
-		expect(env.containerStatuses['target0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
-			ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
-		)
 		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
 		expect(await fsStat('/targets/target0/file0Target.mp4')).toMatchObject({
 			size: 1234,
@@ -113,48 +117,50 @@ describe('Handle unhappy paths', () => {
 			[getLocalTarget('target0', 'file0Target.mp4')]
 		)
 
-		await waitTime(env.WAIT_JOB_TIME)
-		// Expect the Expectation to be waiting:
-		expect(env.expectationStatuses['copy0']).toMatchObject({
-			actualVersionHash: null,
-			statusInfo: {
-				status: /new|waiting/,
-				statusReason: {
-					tech: /not able to access file/i,
+		await waitUntil(() => {
+			// Expect the Expectation to be waiting:
+			expect(env.expectationStatuses['copy0']).toMatchObject({
+				actualVersionHash: null,
+				statusInfo: {
+					status: /new|waiting/,
+					statusReason: {
+						tech: /not able to access file/i,
+					},
 				},
-			},
-		})
+			})
+		}, env.WAIT_JOB_TIME)
 
 		// Now the file can be read from:
 		fs.__mockSetFile('/sources/source0/file0Source.mp4', 1234)
-		await waitTime(env.ERROR_WAIT_TIME)
-		await waitTime(env.WAIT_SCAN_TIME)
 
-		// Expect the Expectation to be waiting:
-		expect(env.expectationStatuses['copy0']).toMatchObject({
-			actualVersionHash: null,
-			statusInfo: {
-				status: 'new',
-				statusReason: {
-					tech: /not able to access target/i,
+		await waitUntil(() => {
+			// Expect the Expectation to be waiting:
+			expect(env.expectationStatuses['copy0']).toMatchObject({
+				actualVersionHash: null,
+				statusInfo: {
+					status: 'new',
+					statusReason: {
+						tech: /not able to access target/i,
+					},
 				},
-			},
-		})
+			})
+		}, env.ERROR_WAIT_TIME + env.WAIT_SCAN_TIME)
 
 		// Now the target can be written to:
 		fs.__mockSetDirectory('/targets/target0', {
 			accessRead: true,
 			accessWrite: true,
 		})
-		await waitTime(env.WAIT_SCAN_TIME)
 
-		// Expect the copy to have completed by now:
+		// Wait until the copy has completed:
+		await waitUntil(() => {
+			expect(env.containerStatuses['target0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
+				ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
+			)
+		}, env.WAIT_JOB_TIME)
 
-		expect(env.containerStatuses['target0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
-			ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
-		)
 		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
 		expect(await fsStat('/targets/target0/file0Target.mp4')).toMatchObject({
 			size: 1234,
@@ -218,22 +224,23 @@ describe('Handle unhappy paths', () => {
 		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('working')
 		expect(listenToCopyFile).toHaveBeenCalledTimes(1)
 
-		await waitTime(env.WORK_TIMEOUT_TIME)
-		await waitTime(env.WAIT_JOB_TIME)
-		// By now, the work should have been aborted, and restarted:
-		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual(expect.stringMatching(/new|waiting/))
+		// Wait until the work have been aborted, and restarted:
+		await waitUntil(() => {
+			expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual(expect.stringMatching(/new|waiting/))
+		}, env.WORK_TIMEOUT_TIME + env.WAIT_JOB_TIME)
 
 		// Add another worker:
 		env.addWorker()
-		await waitTime(env.WAIT_SCAN_TIME)
 
-		// Expect the copy to have completed by now:
-		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
-		expect(env.containerStatuses['target0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
-			ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
-		)
+		// Expect the copy to have completed:
+		await waitUntil(() => {
+			expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
+			expect(env.containerStatuses['target0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
+				ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
+			)
+		}, env.WAIT_SCAN_TIME)
 
 		// Clean up:
 		if (killedWorker) env.removeWorker(killedWorker.id)
@@ -278,22 +285,26 @@ describe('Handle unhappy paths', () => {
 		expect(listenToCopyFile).toHaveBeenCalledTimes(1)
 		expect(hasIntercepted).toBe(1)
 
-		await waitTime(env.WORK_TIMEOUT_TIME)
-		await waitTime(env.WAIT_JOB_TIME)
-		// By now, the work should have been aborted, and restarted:
-		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual(
-			expect.stringMatching(/new|waiting|fulfilled/)
-		)
+		// Wait for the work to be aborted, and restarted:
+		await waitUntil(() => {
+			expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual(
+				expect.stringMatching(/new|waiting|fulfilled/)
+			)
+		}, env.WORK_TIMEOUT_TIME + env.WAIT_JOB_TIME)
 
-		await waitTime(env.WAIT_SCAN_TIME)
+		// Wait for the copy to complete:
+		await waitUntil(() => {
+			expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
+			expect(env.containerStatuses['target0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
+			expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
+				ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
+			)
+		}, env.WAIT_SCAN_TIME)
 
-		// Expect the copy to have completed by now:
-		expect(env.expectationStatuses['copy0'].statusInfo.status).toEqual('fulfilled')
-		expect(env.containerStatuses['target0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0']).toBeTruthy()
-		expect(env.containerStatuses['target0'].packages['package0'].packageStatus?.status).toEqual(
-			ExpectedPackageStatusAPI.PackageContainerPackageStatusStatus.READY
-		)
+		expect(listenToCopyFile).toHaveBeenCalledTimes(1)
+		expect(hasIntercepted).toBe(1)
+
 		// clean up:
 		deferredCallbacks.forEach((cb) => cb())
 	})